cleaning up static data structures
